// front/services/models/webhook.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Global Testing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Webhook
 */
export interface Webhook {
    /**
     * 
     * @type {string}
     * @memberof Webhook
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof Webhook
     */
    'scenario': string;
    /**
     * 
     * @type {string}
     * @memberof Webhook
     */
    'eventType': WebhookEventTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof Webhook
     */
    'url': string;
}

export const WebhookEventTypeEnum = {
    on_failure: 'onFailure',
    on_success: 'onSuccess'
} as const;

export type WebhookEventTypeEnum = typeof WebhookEventTypeEnum[keyof typeof WebhookEventTypeEnum];